MINOR: fix failed StreamsBrokerBounceTest e2e tests (#19839)
Currently some tests in StreamsBrokersBounceTest failed due to error
`The cluster  does not support the STREAMS group protocol or does not
support the  versions of the STREAMS group protocol used by this client
(used  versions: 0 to 0).`

The reason is that under isolated kraft mode, we missed to set both
`unstable.api.versions.enable` and `unstable.feature.versions.enable` to
true to all controllers, which cause `streams.version` fallback to 0 in
the broker side and the above error raise when
StreamsGroupRequestHeartbeat comes to the broker.

This patch add the missing configs to controllers properties if streams
group protocol is used.

Reviewers: Lucas Brutschy <lbrutschy@confluent.io>